+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add support for jsonb and encrypted attributes
+
 ## [0.1.0] - 2023-01-23
 
 - Initial release